CmdJoin: Fix segment fault on join.
Code was attempting to dereference the end () iterator, which is invalid.

Since sets are already ordered, we can use the first and last element directly
instead of using the min/max functions.

Fixes a bug introduced in ( 8ce9e83ab53 "TI-93 #97 Make CLI::getIds() return a
set to guarantee uniqueness of ids" )