#220: Handle escaped quotes in tags database file.
If there are escaped quotes in the tags database file, when the file is read,
then written, additional escape characters were inserted.

For example:

When tag 'a "test"' is written, it gets stored as 'a \"test\"'. Then it's read
back in that form, but the next time it's serialized to disk, the quotes would
be escaped again resulting in 'a \\"test\\"' which has the effect of removing
the escapes for the double-quotes.