Do not allow adding open intervals with a future time
If an interval is started at a future time, as was possible with the
continue command when given a future time, subsequent commands could
fail if the end data was set to the current time.

Which would result in the following message:
  Range.cpp:359: time_t Range::total() const: Assertion `is_open () || end >= start' failed.

This change makes it invalid to start an open interval in the future,
regardless of what command is validating the in the interval.  However
it might still be possible to get in this case if the time on the system
jumps backward after starting an interval.

Reported by @kbcb
Related to #350
Closes #364

Signed-off-by: Shaun Ruffell <sruffell@sruffell.net>