Add defines and bump libshared peg so faster timegm() is used via cmake.h
This fix will cause cmake to detect and define HAVE_TIMEGM in cmake.h if
the system libc has a timegm() function, and updates the libshared peg
to incorporate cmake.h in the included libshared build.  This way, the
slow Datetime::timegm() implementation is not used, but rather the fast
one in libc.  This fast one depends on HAVE_TIMEGM being defined by the
function in libshared's Datetime.cpp.

Fixes #703 which was a significant performance regression in 1.9.0
release.

Signed-off-by: Scott Mcdermott <scott@smemsh.net>