Pass color settings to the tables
The builders of the tags and summary table did not pass the value of the color setting.
Also, the 'Table' component had an internal TTY check which could interfere with the color setting of the main application.

Add test for color hint
Reapply command line hints after loading config files

Closes #727

Signed-off-by: Thomas Lauf <thomas.lauf@tngtech.com>